+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - **Fixed a `Table` bug where filter icon is hidden when `manualFilters` is on and no data is returned after filtering.**
 
+### What's new
+
+- **Added `Fieldset` component** - basic and disabled.
+
 ## [1.14.1]
 
 `2021-08-05`